Add files via upload
I coded a very simple game where the computer picks a random number between 1 and 10 and the operator have to guess the number.  I had to learn about srand() which sets the seed which is used by rand to generate “random” numbers. It took me a good amount of time to get the random number generator going and the loop to function properly but had lots of fun!.